Move offline URL helper functions to home.utils
